// types/enums.ts
export enum Trend {
  INSUFICIENT = 0,
  DISAPPEARED = 1,
  CONSTANT_FALL = 2,
  STABLE = 3,
  GROWING = 4,
  SOARED = 5,
}

export enum AbcCurve {
  C = 1,
  B = 2,
  A = 3,
}

export enum AbcCurveChange {
  FELL = 0,
  SAME = 1,
  GREW = 2,
}

export enum Momentum {
  FALLING = 0,
  STABLE = 1,
  GROWING = 2,
}